유저í프로필조회 수정

+++ src/main/java/cocktail/dto/UserProfileDto.java
@@ -16,6 +16,6 @@ public class UserProfileDto {
     private boolean loginUser;
     private boolean follow;
     private UserDto userDto;
-    private int userFollowerCount;
-    private int userFollowingCount;
+    private long userFollowerCount;
+    private long userFollowingCount;
 }
